Update trackbars.py
Changes made will remain on the frame, added more colors, created switches to toggle shape coordinates, area, fill, and dimensions on and off, change their thickness. Added feature to toggle inverse of the HSV mask created. (need to modify this program further so that 2 instances of the trackbar can coexist and make changes on the same frame)